fix: roboto mono, code view storybook sidenav
