+++ src/components/reusable/codeView/blockCodeView/blockCodeView.scss
@@ -78,7 +78,7 @@ $border-radius: 4px;
             code {
                 display: block;
                 white-space: inherit;
-                font-family: monospace;
+                font-family: var(--kd-font-family-code-view);
                 font-size: 14px;
                 line-height: 1.5;
                 text-align: left;
