+++ src/components/reusable/codeView/inlineCodeView/inlineCodeView.scss
@@ -10,13 +10,12 @@ $default-light-font-color: #3D3C3C;
 }
 
 code {
-    font-family: monospace;
     line-height: 1.5;
     word-break: break-word;
     vertical-align: text-bottom;
     margin: 0 4px;
     padding: 4px 8px;
-    font-family: monospace;
+    font-family: var(--kd-font-family-code-view);
     font-size: max(calc(var(--inline-snippet-font-size, 14) * 1px), 14px);
     border-radius: 2px;
