+++ tsconfig.json
@@ -36,6 +36,10 @@
       }
     ]
   },
-  "include": ["src/**/*.ts", "declarations.d.ts"],
+  "include": [
+    "src/**/*.ts",
+    "declarations.d.ts",
+    "src/components/reusable/codeView/blockCodeView.stories.js"
+  ],
   "exclude": ["src/**/*.test.ts"]
 }
